Change project architecture & add test project
