+++ Simple.Http/Simple.Http.Tests/UnitTest1.cs
@@ -1,0 +1,10 @@
+﻿namespace Simple.Http.Tests;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+
+    }
+}
