--- Simple.Http/Simple.Http.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <Version>1.1.0</Version>
-    <Authors>ThomasBernard</Authors>
-    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <PackageIconUrl>https://raw.githubusercontent.com/ThomasBernard03/Simple.Http/dev/simple.png</PackageIconUrl>
-    <PackageProjectUrl>https://github.com/ThomasBernard03/Simple.Http</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/ThomasBernard03/Simple.Http</RepositoryUrl>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-  </ItemGroup>
-
-</Project>
