+++ Simple.Http/Simple.Http.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32014.148
+# Visual Studio Version 16
+VisualStudioVersion = 25.0.1704.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Simple.Http", "Simple.Http.csproj", "{5D44C841-8F41-411E-8783-3F30C503BD29}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Simple.Http", "Simple.Http\Simple.Http.csproj", "{361B8460-8F12-45F8-ADCD-C3591DECE8EC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Simple.Http.Tests", "Simple.Http.Tests\Simple.Http.Tests.csproj", "{8E59A1B5-3CF0-4148-8B0A-26DF6C1C6F2D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +13,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5D44C841-8F41-411E-8783-3F30C503BD29}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5D44C841-8F41-411E-8783-3F30C503BD29}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5D44C841-8F41-411E-8783-3F30C503BD29}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5D44C841-8F41-411E-8783-3F30C503BD29}.Release|Any CPU.Build.0 = Release|Any CPU
+		{361B8460-8F12-45F8-ADCD-C3591DECE8EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{361B8460-8F12-45F8-ADCD-C3591DECE8EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{361B8460-8F12-45F8-ADCD-C3591DECE8EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{361B8460-8F12-45F8-ADCD-C3591DECE8EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E59A1B5-3CF0-4148-8B0A-26DF6C1C6F2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E59A1B5-3CF0-4148-8B0A-26DF6C1C6F2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E59A1B5-3CF0-4148-8B0A-26DF6C1C6F2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E59A1B5-3CF0-4148-8B0A-26DF6C1C6F2D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FC9DD94D-F66B-4CC3-92AC-385136880812}
+		SolutionGuid = {DD98ED04-21EB-46B2-84AE-6A3427173F40}
 	EndGlobalSection
 EndGlobal
